Set an image in the docker-compose file.
Apparently, commenting out the image makes the whole thing completely
broken. This is the cause of all of our recent misery.

As a temporary measure, I've just set it to mvbachman/dms, but you'll
want to change this to whatever your dockerhub or GitHub Container
Registry username is once you've set one up and pushed this image to
that registry.