Bind volumes to allow hot reloads.
Before this change, the entire docker container would have to be rebuilt
in order for any of the changes you make to propagate into the
container. This isn't optimal if you want to test the container locally,
so I've set up volumes for all the business logic to enable hot reloads.

Now, as soon as you change something in one of these folders (like
changing the settings.py or altering the code of one of your views), the
gunicorn server will detect the change and restart the server
immediately. That way, you'll be able to see your change in the browser
as soon as you refresh, just like when running a normal django runserver