Add a way to retry to do run migrations
This is so if the migrations fail because the database is still not
ready then it will retry until it is ready. After trying then times it
will just kill the process and the server won't be allowed to run. If
migrations are not run this could cause potential problems in production.