Remove DATA_UPLOAD_MAX_NUMBER_FIELDS so it uses default
The DATA_UPLOAD_MAX_NUMBER_FIELDS is used to specify the number of
parameters that may be received via GET or POST before the TooManyFields
error is raised. The default is 1000 which is plenty and the only reason
I had changed this before was to try and delete a bunch of records at
once using the admin panel. There was a better way to achieve this(using
shell) and forgot to remove this. 1000 parameters should be plenty.
Django doc entry for it: https://bit.ly/2Iwn2ic