Run celery as a service in the docker container
Before I tried simply running celery in the dms service which is what
runs the server but this fails miserably. The server never gets ran
because celery is running and that line never gets executed. This
service is simply a copy of the dms service with a different entrypoint,
port, and name. Also added it to the ci compose so that can be tested in
the pipeline.