Add Clean to EmployeeCreationForm
Added a check to make sure the employee_id is not already in use. Before
if an attempt was made to add an employee with an ID that existed no
errors would be added to the form but the save method in the form would
raise a key already exists.