Edit create_settlement_document
When a paragraph was less more than 175 in width it would split it up
then draw it 1.125in form the left side of the document. However, when
it was not it would draw it 1.75in from the left of the document. Not
really sure why it was like this so I changed it for it to be the same
distance from the left.