Edit Settlement model
Changed Settlement to have a OneToMany relationship with employee as
opposed to a OneToOne relationship. The reason being that an employee
could have multiple Settlement or simply if someone deletes a Settlement
it only sets it inactive and never actually deletes it. Which would of
course cause issues when making a new one.