Edit start command to run local server on debug
The start command will now check if debug mode is turned on or not, if
it is it will simply use the runserver command otherwise it will use
gunicorn to run the server. This will allow for simpler testing when
in development to support hot swaps and see changes happen immediately.