Edit traefik config
As of now I do not think I will be using the dashboard so turning it off
to not have any vulnerabilities when in production. Also changing the
log level to be an environment variable for anytime there are issues and
testing needs to be done. If the env variable is not set it will default
to INFO.