Add comparison to existing groups
Before I was setting it to create the necessary groups and simply log if
it could not make a group or if it made it. However, if the group
already existed it would log it saying it could not create. I made it
easier to identify now to log existing groups as well.